Move registration cases to own folder
[stage-19]
