+++ test/e2e/apps.scenarios.js
@@ -6,4 +6,4 @@ require('./common/common.cases.js');
 require('./template-editor/template-editor.cases');
 require('./template-editor/template-editor2.cases');
 require('./common-header/common-header.cases');
-require('./common-header/registration.cases');
+require('./registration/registration.cases');
