+++ test/e2e/editor/cases/professional-widgets.js
@@ -7,7 +7,7 @@ var PresentationsListPage = require('./../pages/presentationListPage.js');
 var WorkspacePage = require('./../pages/workspacePage.js');
 var PlaceholderPlaylistPage = require('./../pages/placeholderPlaylistPage.js');
 var StoreProductsModalPage = require('./../pages/storeProductsModalPage.js');
-var PricingComponentModalPage = require('./../../common/pages/pricingComponentModalPage.js');
+var PricingComponentModalPage = require('./../../registration/pages/pricingComponentModalPage.js');
 var PresentationItemModalPage = require('./../pages/presentationItemModalPage.js');
 var PresentationModalPage = require('./../../schedules/pages/presentationModalPage.js');
 var TwitterSettingsPage = require('./../pages/twitterSettingsPage.js');
