+++ test/e2e/registration/cases/checkout.js
@@ -5,11 +5,11 @@
   var expect = require('rv-common-e2e').expect;
   var assert = require('rv-common-e2e').assert;
   var helper = require('rv-common-e2e').helper;
-  var CommonHeaderPage = require('./../pages/commonHeaderPage.js');
-  var HomePage = require('./../pages/homepage.js');
+  var CommonHeaderPage = require('./../../common-header/pages/commonHeaderPage.js');
+  var HomePage = require('./../../common-header/pages/homepage.js');
   var SignInPage = require('./../../common/pages/signInPage.js');
-  var PurchaseFlowModalPage = require('./../../common/pages/purchaseFlowModalPage.js');
-  var PricingComponentModalPage = require('./../../common/pages/pricingComponentModalPage.js');
+  var PurchaseFlowModalPage = require('./../pages/purchaseFlowModalPage.js');
+  var PricingComponentModalPage = require('./../pages/pricingComponentModalPage.js');
 
   var Checkout = function() {
