+++ test/e2e/registration/cases/registration.js
@@ -5,7 +5,7 @@
   var expect = require('rv-common-e2e').expect;
   var assert = require('rv-common-e2e').assert;
   var helper = require('rv-common-e2e').helper;
-  var CommonHeaderPage = require('./../pages/commonHeaderPage.js');
+  var CommonHeaderPage = require('./../../common-header/pages/commonHeaderPage.js');
   var RegistrationModalPage = require('./../pages/registrationModalPage.js');
   var SignInPage = require('./../../common/pages/signInPage.js');
   var SignUpPage = require('./../../common/pages/signUpPage.js');
