+++ test/e2e/registration/pages/pricingComponentModalPage.js
