+++ test/e2e/registration/pages/purchaseFlowModalPage.js
