+++ test/e2e/registration/pages/registrationModalPage.js
