+++ test/e2e/registration/utils/mailListener.js
